requirements.txt
django==4.2.7
lxml==4.9.3
pyquery==2.0.0
gunicorn==20.1.0
psycopg2-binary==2.8.5
coverage==7.2.7
factory_boy==3.2.1
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
django-webtest==1.9.11